backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
celery==5.3.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
tenacity==8.2.3
scrapy==2.11.1
selenium==4.18.1
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.8.2
tenacity==8.2.3
 APScheduler==3.10.4
